Refactor Levels QML implementation
